@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --webpack",
-    "build": "next build --webpack",
+    "dev": "next dev",
+    "build": "next build",
     "start": "next start",
     "lint": "biome lint .",
     "lint:fix": "biome lint --write .",
@@ -19,7 +19,6 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@hookform/resolvers": "^3.3.2",
     "@mui/material": "^5.15.20",
     "@mui/x-charts": "^7.7.1",
     "@neondatabase/serverless": "^1.0.2",
@@ -30,7 +29,6 @@
     "@radix-ui/react-context-menu": "^2.1.5",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-progress": "^1.1.0",
     "@radix-ui/react-scroll-area": "^1.0.5",
     "@radix-ui/react-select": "^2.1.1",
     "@radix-ui/react-separator": "^1.0.3",
@@ -42,10 +40,7 @@
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.1.2",
     "@shadcn/ui": "^0.0.4",
-    "@tanstack/react-query": "^5.17.15",
     "@tanstack/react-table": "^8.20.1",
-    "@types/md5": "^2.3.5",
-    "@types/pdf-parse": "^1.1.4",
     "axios": "^1.6.7",
     "better-auth": "^1.4.18",
     "class-variance-authority": "^0.7.0",
@@ -53,18 +48,14 @@
     "drizzle-orm": "^0.40.0",
     "embla-carousel-react": "^8.0.0-rc22",
     "framer-motion": "^11.3.21",
-    "fs": "^0.0.1-security",
-    "init": "^0.1.2",
     "langchain": "^0.1.4",
     "lucide-react": "^0.294.0",
     "marked": "^13.0.3",
-    "md5": "^2.3.0",
     "nanoid": "^5.0.4",
-    "next": "16.0.7",
+    "next": "16.1.6",
     "next-themes": "^0.2.1",
     "openai": "^4.6.0",
-    "pdf-parse": "^1.1.1",
-    "pdfjs-dist": "^4.10.38",
+    "pdf-parse": "1.1.1",
     "pg": "^8.13.3",
     "react": "^18",
     "react-audio-player": "^0.17.0",
@@ -79,8 +70,7 @@
     "sonner": "^1.4.41",
     "tailwind-merge": "^2.1.0",
     "tailwindcss-animate": "^1.0.7",
-    "uuid": "^10.0.0",
-    "zod": "^3.22.4"
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
